portability fix for kaffe-qt-awt
2005-09-16  Dalibor Topic  <robilad@kaffe.org>

        * libraries/clib/awt/qt/fnt.cc:
        (Java_java_awt_Toolkit_fntBytesWidth)
        Cast jb to uintptr_t for portable pointer arithmetic.

        Suggested by:  Gwenole Beauchesne <gbeauchesne@mandriva.com>

+++ libraries/clib/awt/qt/fnt.cc
@@ -14,6 +14,8 @@
 #include <qfont.h>
 #include <qfontmetrics.h>
 
+#include "config.h"
+#include "config-int.h"
 #include "toolkit.h"
 
 /**
@@ -187,7 +189,7 @@ Java_java_awt_Toolkit_fntBytesWidth ( JNIEnv* env, jclass clazz,
   if ( off  + len > n ) 
     len = n - off;
   QByteArray a;
-  a.setRawData((const char*)((long)jb + off), len);
+  a.setRawData((const char*)((uintptr_t)jb + off), len);
   w = fm.width(QString(a),len);
   env->ReleaseByteArrayElements( jBytes, jb, JNI_ABORT);
   DBG(AWT_FNT, qqDebug("fntBytesWidth %s %d\n",jb + off, w));
